pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = ""
authors = ["Rovshen Tagangylyjov <rovshentagangylyjov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
django-cors-headers = "^3.10.1"
drf-yasg = "^1.20.0"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
